Table Column Name 수정에 따른 코드 수정

+++ src/main/java/com/investory/market/presentation/dto/response/StockPriceResponse.java
@@ -15,7 +15,7 @@ public record StockPriceResponse(
         Long closePrice,
         BigDecimal dailyReturnRate,
         Long tradingVolume,
-        Long tradingValue
+        BigDecimal tradingValue
 ) {
     public static StockPriceResponse from(StockPrice stockPrice) {
         return new StockPriceResponse(
